ADD: 컴포넌트 파일 세분화
