+++ src/pages/Main/Main.scss
@@ -1,104 +1,8 @@
-@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR&display=swap');
 * {
-  font-family: 'Noto Sans KR', sans-serif;
   font-size: 18px;
 }
+
 .main {
   display: flex;
   flex-direction: column;
-}
-.promotion {
-  background-color: rgb(244, 244, 244);
-  .promotionPhrase {
-    margin: 40px;
-    font-size: 40px;
-    font-weight: bold;
-  }
-  .promotionBtn {
-    margin-left: 40px;
-    margin-bottom: 40px;
-    width: 200px;
-    height: 50px;
-    border: 0;
-    border-radius: 5px;
-    background-color: #201c1d;
-    color: white;
-    font-size: 18px;
-  }
-}
-.delivery {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  .deliveryLayout {
-    display: flex;
-    flex-direction: row;
-    .phraseLayout {
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      text-align: center;
-      .deliveryWord {
-        font-size: 50px;
-        font-weight: bold;
-      }
-      .deliveryPhrase {
-        margin-top: 10px;
-      }
-    }
-    .deliveryIcon {
-      width: 100px;
-      margin-left: 60px;
-    }
-  }
-}
-
-.BestProductTitle {
-  margin-top: 100px;
-  margin-right: auto;
-  margin-left: auto;
-  font-size: 32px;
-  font-weight: bold;
-}
-.BestProductList {
-  display: grid;
-  grid-template-columns: 400px 400px 400px;
-  grid-template-rows: 400px 400px;
-  grid-gap: 30px;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  margin: 50px 50px 100px 50px;
-  font-size: 32px;
-  .BestProductItem {
-    margin-top: 40px;
-    background-color: rgb(244, 244, 244);
-    width: 400px;
-    height: 350px;
-    .productImage {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      margin: auto;
-      padding-top: 20px;
-      width: 250px;
-    }
-  }
-  .productInformation {
-    display: flex;
-    flex-direction: column;
-    margin-top: 10px;
-    text-align: left;
-    float: left;
-    font-weight: bold;
-    font-size: 18px;
-    .productPriceInformation {
-      font-weight: 400;
-      font-size: 16px;
-
-      color: gray;
-    }
-  }
 }
