+++ src/pages/Main/Section/Promotion/Promotion.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import './Promotion.scss';
 
 const Promotion = () => {
   return (
