+++ src/pages/Main/Section/Promotion/Promotion.scss
@@ -1,0 +1,19 @@
+.promotion {
+  background-color: rgb(244, 244, 244);
+  .promotionPhrase {
+    margin: 40px;
+    font-size: 40px;
+    font-weight: bold;
+  }
+  .promotionBtn {
+    margin-left: 40px;
+    margin-bottom: 40px;
+    width: 200px;
+    height: 50px;
+    border: 0;
+    border-radius: 5px;
+    background-color: rgb(11, 10, 10);
+    color: white;
+    font-size: 18px;
+  }
+}
